+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
-## [1.1.1] - XXXX-XX-XX
+## [1.1.1] - 2023-07-21
 
 ### Fixed
 
 - Image now have attribute applied properly (check Readme for placement)
 - Fixed some naming
+- Fixed handling direct Assets path in images
+
+### Added
+
+- Now handle packages paths
+- Added package: search special links
 
 ## [1.1.0] - 2023-06-30